# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sales Analytics & Visualization Layer API - Odoo CRM integration
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LoginBody {
  username: string;
  password: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  BD: "BD",
  TL: "TL",
  TS: "TS",
  ADMIN: "ADMIN",
} as const;

export interface User {
  id: number;
  username: string;
  name: string;
  role: UserRole;
  managerId?: number | null;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export type CreateUserBodyRole =
  (typeof CreateUserBodyRole)[keyof typeof CreateUserBodyRole];

export const CreateUserBodyRole = {
  BD: "BD",
  TL: "TL",
  TS: "TS",
  ADMIN: "ADMIN",
} as const;

export interface CreateUserBody {
  username: string;
  password: string;
  name: string;
  role: CreateUserBodyRole;
  managerId?: number | null;
}

export interface Ticket {
  id: number;
  odooId: number;
  clientName: string;
  phone?: string | null;
  email?: string | null;
  expectedRevenue?: number | null;
  stage: string;
  probability?: number | null;
  salesperson?: string | null;
  lastUpdate?: string | null;
  createDate?: string | null;
  assignedUserId?: number | null;
  tags: string[];
  internalNote?: string | null;
  syncedAt: string;
}

export type ActivityType = (typeof ActivityType)[keyof typeof ActivityType];

export const ActivityType = {
  call: "call",
  meeting: "meeting",
  whatsapp: "whatsapp",
  email: "email",
  note: "note",
} as const;

export interface Activity {
  id: number;
  ticketId: number;
  userId?: number | null;
  type: ActivityType;
  summary?: string | null;
  internalOnly: boolean;
  doneAt: string;
  createdAt: string;
}

export type TicketDetail = Ticket & {
  assignedUser?: User | null;
  activities?: Activity[];
};

export interface TicketListResponse {
  tickets: Ticket[];
  total: number;
  limit: number;
  offset: number;
}

export interface UpdateTicketBody {
  assignedUserId?: number | null;
  tags?: string[];
  internalNote?: string | null;
}

export interface SyncResult {
  success: boolean;
  created: number;
  updated: number;
  errors: number;
  message: string;
  syncedAt: string;
}

export type SyncLogStatus = (typeof SyncLogStatus)[keyof typeof SyncLogStatus];

export const SyncLogStatus = {
  pending: "pending",
  success: "success",
  error: "error",
} as const;

export interface SyncLog {
  id: number;
  startedAt: string;
  finishedAt?: string | null;
  status: SyncLogStatus;
  created: number;
  updated: number;
  errors: number;
  message?: string | null;
}

export interface DashboardSummary {
  totalPipelineValue: number;
  weightedRevenue: number;
  totalTickets: number;
  openTickets: number;
  wonTickets: number;
  lostTickets: number;
  avgProbability: number;
  conversionRate: number;
}

export interface StageStats {
  stage: string;
  count: number;
  totalRevenue: number;
  avgProbability: number;
}

export interface UserPerformance {
  userId?: number | null;
  userName: string;
  role?: string | null;
  totalDeals: number;
  totalRevenue: number;
  weightedRevenue: number;
  wonDeals: number;
  conversionRate: number;
}

export interface ServiceRevenue {
  service: string;
  totalRevenue: number;
  count: number;
}

export interface ConversionStage {
  stage: string;
  count: number;
  dropOffRate: number;
  conversionRate: number;
}

export type CreateActivityBodyType =
  (typeof CreateActivityBodyType)[keyof typeof CreateActivityBodyType];

export const CreateActivityBodyType = {
  call: "call",
  meeting: "meeting",
  whatsapp: "whatsapp",
  email: "email",
  note: "note",
} as const;

export interface CreateActivityBody {
  ticketId: number;
  type: CreateActivityBodyType;
  summary?: string | null;
  doneAt: string;
}

export interface OdooMapping {
  id: number;
  odooSalesperson: string;
  localUserId: number;
  localUser?: User | null;
  createdAt: string;
}

export interface CreateMappingBody {
  odooSalesperson: string;
  localUserId: number;
}

export type ListTicketsParams = {
  stage?: string;
  userId?: number;
  teamRole?: ListTicketsTeamRole;
  service?: string;
  minProbability?: number;
  search?: string;
  limit?: number;
  offset?: number;
};

export type ListTicketsTeamRole =
  (typeof ListTicketsTeamRole)[keyof typeof ListTicketsTeamRole];

export const ListTicketsTeamRole = {
  BD: "BD",
  TL: "TL",
  TS: "TS",
} as const;

export type GetDashboardSummaryParams = {
  userId?: number;
};

export type GetStageDistributionParams = {
  userId?: number;
};

export type GetPerformanceByUserParams = {
  managerId?: number;
};

export type ListActivitiesParams = {
  ticketId?: number;
  userId?: number;
};