# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Sales Analytics & Visualization Layer API - Odoo CRM integration
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with username and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /tickets:
    get:
      operationId: listTickets
      tags: [tickets]
      summary: List all tickets from cached DB
      security:
        - bearerAuth: []
      parameters:
        - name: stage
          in: query
          schema:
            type: string
        - name: userId
          in: query
          schema:
            type: integer
        - name: teamRole
          in: query
          schema:
            type: string
            enum: [BD, TL, TS]
        - name: service
          in: query
          schema:
            type: string
        - name: minProbability
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Paginated ticket list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TicketListResponse"
  /tickets/{id}:
    get:
      operationId: getTicket
      tags: [tickets]
      summary: Get a ticket by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Ticket details with activities and notes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TicketDetail"
    patch:
      operationId: updateTicket
      tags: [tickets]
      summary: Update local metadata for a ticket (tags, notes, assignedUserId)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTicketBody"
      responses:
        "200":
          description: Updated ticket
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
...
  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Get overall pipeline summary (total value, weighted revenue, counts)
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Dashboard summary metrics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
  /dashboard/stage-distribution:
    get:
      operationId: getStageDistribution
      tags: [dashboard]
      summary: Get ticket counts and values grouped by pipeline stage
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Stage distribution data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StageStats"
  /dashboard/performance-by-user:
    get:
      operationId: getPerformanceByUser
      tags: [dashboard]
      summary: Get performance metrics grouped by salesperson/user
      security:
        - bearerAuth: []
      parameters:
        - name: managerId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Per-user performance metrics
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserPerformance"
  /dashboard/revenue-by-service:
    get:
      operationId: getRevenueByService
      tags: [dashboard]
      summary: Get expected revenue grouped by service tag
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Revenue by service
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ServiceRevenue"
  /dashboard/conversion-funnel:
    get:
      operationId: getConversionFunnel
      tags: [dashboard]
      summary: Get conversion rates through each pipeline stage
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Conversion funnel data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ConversionStage"
  /activities:
    get:
      operationId: listActivities
      tags: [activities]
      summary: List local activities for a ticket
      security:
        - bearerAuth: []
      parameters:
        - name: ticketId
          in: query
          schema:
            type: integer
        - name: userId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of activities
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Activity"
    post:
      operationId: createActivity
      tags: [activities]
      summary: Create a local activity (internal only, not synced to Odoo)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateActivityBody"
      responses:
        "201":
          description: Activity created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Activity"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    LoginBody:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required:
        - token
        - user
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [BD, TL, TS, ADMIN]
        managerId:
          type: integer
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - username
        - name
        - role
        - createdAt
...
    Ticket:
      type: object
      properties:
        id:
          type: integer
        odooId:
          type: integer
        clientName:
          type: string
        phone:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        expectedRevenue:
          type: number
          nullable: true
        stage:
          type: string
        probability:
          type: number
          nullable: true
        salesperson:
          type: string
          nullable: true
        lastUpdate:
          type: string
          format: date-time
          nullable: true
        createDate:
          type: string
          format: date-time
          nullable: true
        assignedUserId:
          type: integer
          nullable: true
        tags:
          type: array
          items:
            type: string
        internalNote:
          type: string
          nullable: true
        syncedAt:
          type: string
          format: date-time
      required:
        - id
        - odooId
        - clientName
        - stage
        - tags
        - syncedAt
    TicketDetail:
      allOf:
        - $ref: "#/components/schemas/Ticket"
        - type: object
          properties:
            assignedUser:
              $ref: "#/components/schemas/User"
              nullable: true
            activities:
              type: array
              items:
                $ref: "#/components/schemas/Activity"
    TicketListResponse:
      type: object
      properties:
        tickets:
          type: array
          items:
            $ref: "#/components/schemas/Ticket"
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
      required:
        - tickets
        - total
        - limit
        - offset
    UpdateTicketBody:
      type: object
      properties:
        assignedUserId:
          type: integer
          nullable: true
        tags:
          type: array
          items:
            type: string
        internalNote:
          type: string
          nullable: true
...
    DashboardSummary:
      type: object
      properties:
        totalPipelineValue:
          type: number
        weightedRevenue:
          type: number
        totalTickets:
          type: integer
        openTickets:
          type: integer
        wonTickets:
          type: integer
        lostTickets:
          type: integer
        avgProbability:
          type: number
        conversionRate:
          type: number
      required:
        - totalPipelineValue
        - weightedRevenue
        - totalTickets
        - openTickets
        - wonTickets
        - lostTickets
        - avgProbability
        - conversionRate
    StageStats:
      type: object
      properties:
        stage:
          type: string
        count:
          type: integer
        totalRevenue:
          type: number
        avgProbability:
          type: number
      required:
        - stage
        - count
        - totalRevenue
        - avgProbability
    UserPerformance:
      type: object
      properties:
        userId:
          type: integer
          nullable: true
        userName:
          type: string
        role:
          type: string
          nullable: true
        totalDeals:
          type: integer
        totalRevenue:
          type: number
        weightedRevenue:
          type: number
        wonDeals:
          type: integer
        conversionRate:
          type: number
      required:
        - userName
        - totalDeals
        - totalRevenue
        - weightedRevenue
        - wonDeals
        - conversionRate
    ServiceRevenue:
      type: object
      properties:
        service:
          type: string
        totalRevenue:
          type: number
        count:
          type: integer
      required:
        - service
        - totalRevenue
        - count
    ConversionStage:
      type: object
      properties:
        stage:
          type: string
        count:
          type: integer
        dropOffRate:
          type: number
        conversionRate:
          type: number
      required:
        - stage
        - count
        - dropOffRate
        - conversionRate
    Activity:
      type: object
      properties:
        id:
          type: integer
        ticketId:
          type: integer
        userId:
          type: integer
          nullable: true
        type:
          type: string
          enum: [call, meeting, whatsapp, email, note]
        summary:
          type: string
          nullable: true
        internalOnly:
          type: boolean
        doneAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - ticketId
        - type
        - internalOnly
        - doneAt
        - createdAt
    CreateActivityBody:
      type: object
      properties:
        ticketId:
          type: integer
        type:
          type: string
          enum: [call, meeting, whatsapp, email, note]
        summary:
          type: string
          nullable: true
        doneAt:
          type: string
          format: date-time
      required:
        - ticketId
        - type
        - doneAt
...